% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{calc_ndvi_buff}
\alias{calc_ndvi_buff}
\title{Calculate normalized difference vegetation index (NDVI) within a buffer around a point location}
\usage{
calc_ndvi_buff(
  lon = -1.469,
  lat = 51.7779,
  dist1 = 1000,
  dist2 = 50,
  start_date = "2019-01-01",
  end_date = "2022-12-31",
  cloud_cover = 10,
  tf_id = 85
)
}
\arguments{
\item{dist1}{A numeric indicating the buffer distance in meters around the point of interest. Default is 1000.}

\item{dist2}{A numeric indicating the buffer distance in meters around the point used to filter the image collection. Default is 50.}

\item{start_date}{A character indicating the start date of the image collection in the format 'YYYY-MM-DD'. Default is '2021-01-01'.}

\item{end_date}{A character indicating the end date of the image collection in the format 'YYYY-MM-DD'. Default is '2021-12-31'.}

\item{cloud_cover}{A numeric indicating the maximum percentage of cloudy pixels allowed in the image collection. Default is 10.}

\item{tf_id}{An optional#' parameter indicating an ID or name for the point of interest. Default is NULL.}
}
\value{
A list object containing the following components:
map: A map with two layers displaying the NDVI and the input point location.
tidy_dates: A tidied version of the image collection with start and end dates associated with each image in the collection#' image_ids: A vector of image IDs corresponding to the images used in the calculation.
raster: A raster layer object of the NDVI values within the buffer region.
ndvi_stats: Summary statistics of the NDVI values within the buffer region, including median, minimum, maximum, mean and standard deviation.
}
\description{
This function takes in a latitude, longitude, buffer distance, start and end time period, and cloud cover percentage,
and returns a map of NDVI within the buffer region, along with statistics on the NDVI values within the buffer.
}
\examples{

# Calculate NDVI and its summary statistics within a buffer and a point of interest for default parameters
calc_ndvi_buff()

}
\references{
\itemize{
\item COPERNICUS/S2_SR_HARMONIZED image collection.
\item rgee package documentation: https://csaybar.github.io/rgee/index.html
\item tidyrgee package documentation: https://csaybar.github.io/tidyrgee/index.html
\item zoo package documentation: https://cran.r-project.org/web/packages/zoo/index.html
\item terra package documentation: https://rspatial.org/terr'
}
}
\seealso{
\code{\link{as_tidyee}}, \code{\link{ee_extract_tidy}}
}
