% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_ndvi.R
\name{extract_ndvi}
\alias{extract_ndvi}
\title{Extracts NDVI from Google Earth Engine}
\usage{
extract_ndvi(lat, lon, start_date, end_date, x)
}
\arguments{
\item{lat}{numeric Latitude of the region of interest}

\item{lon}{numeric Longitude of the region of interest}

\item{start_date}{character Start date of the time series in yyyy-mm-dd format}

\item{end_date}{character End date of the time series in yyyy-mm-dd format}

\item{x}{character User-defined identifier used to label the output file}
}
\description{
Extracts NDVI from Google Earth Engine
}
