% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dw_landcover.R
\name{get_dw_landcover}
\alias{get_dw_landcover}
\title{Get land cover data from the Dynamic World Land Cover dataset for a given location}
\usage{
get_dw_landcover(
  tf = tf_id,
  lon = 0.0969,
  lat = 51.578,
  dist = 500,
  start_date = "2022-01-01",
  end_date = "2022-12-31"
)
}
\arguments{
\item{tf}{text: specify which TerraFusion site ID to use, defaults to tf_id}

\item{lon}{numeric: longitude of the center of the area of interest, defaults to 0.0969}

\item{lat}{numeric: latitude of the center of the area of interest, defaults to 51.578}

\item{dist}{numeric: distance around the center point to create a circular buffer in meters, defaults to 500}

\item{start_date}{character: starting date of the image collection, defaults to '2022-01-01'}

\item{end_date}{character: ending date of the image collection, defaults to '2022-12-31'}
}
\value{
list: a list containing the following items:
\item{map}{a map object showing the image collection and area of interest}
\item{image_dates}{vector of image dates in the collection}
\item{#' image_ids}{vector of image IDs in the collection}
\item{raster}{a cropped raster object showing the land cover statistics for the area of interest}
\item{lc_stats}{a data frame containing the land cover class counts for the area of interest}
\item{dw_df}{a data frame containing all the land cover data for the area of interest}
\item{tf_id}{the TerraFusion site ID used}
\item{buffer}{the buffer distance used to select the area of interest}
}
\description{
Downloads the image collection from the Dynamic World Land Cover (DW) dataset on Google Earth Engine (GEE) platform, filters it using a given location and time frame, and returns the land cover statistics for the given area.
}
