% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndvi_ts.R
\name{get_ndvi_time_series}
\alias{get_ndvi_time_series}
\title{Get NDVI time series}
\usage{
get_ndvi_time_series(
  lon,
  lat,
  start = "2019-01-01",
  end = "2023-01-01",
  ic = "COPERNICUS/S2_SR",
  buffer1 = 50,
  buffer2 = 1000
)
}
\arguments{
\item{lon}{numeric: Longitude of the location of interest.}

\item{lat}{numeric: Latitude of the location of interest.}

\item{start}{character: Start date of the time frame, in YYYY-MM-DD format (default is '2019-01-01').}

\item{end}{character: End date of the time frame, in YYYY-MM-DD format (default is '2023-01-01').}

\item{ic}{character: Google Earth Engine ImageCollection name (default is "COPERNICUS/S2_SR").}

\item{buffer1}{numeric: Buffer around point of interest, in meters (default is 100).}

\item{buffer2}{numeric: Buffer around point of interest to retrieve summary statistics, in meters (default is 1000).}
}
\value{
list containing:
\describe{
\item{summary_point}{Summary statistics of NDVI values at the specified point#'  \item{sd}{Standard deviation of NDVI values at the specified point}
\item{summary_buff}{Summary statistics of NDVI values within buffer2 around the specified point}
\item{sd_buff}{Standard deviation of NDVI values within buffer2 around the specified point}
\item{plot}{Visualization of the time series of NDVI values at the specified point}
}
}
\description{
This function retrieves the time series of Normalized Difference Vegetation Index (NDVI) values
at a specified location and within a given time frame. The function returns a list containing the
summary statistics of NDVI values at the specified point and a visualization of the time series.
}
\examples{
get_ndvi_time_series(-121.7, 36.9)

}
