% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbn_species_counts.R
\name{get_species_counts}
\alias{get_species_counts}
\title{This function takes a longitude, latitude, taxa, and optional search radius and returns a data frame with the species, count, and other information for records found on the NBN Atlas.}
\usage{
get_species_counts(lon, lat, taxa, radius = 1, pageSize = 1000)
}
\arguments{
\item{lon}{Longitude of centre point for search}

\item{lat}{Latitude of centre point for search}

\item{taxa}{Scientific or common name of taxa to search for}

\item{radius}{Search radius in kilometres around the centre point, defaults to 1 km}
}
\value{
A data frame of species counts with columns for scientific name, common name, count, and other information.
}
\description{
This function takes a longitude, latitude, taxa, and optional search radius and returns a data frame with the species, count, and other information for records found on the NBN Atlas.
}
\examples{
get_species_counts( -1.394457, 50.936542, "Birds" )

}
