% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_data.R
\name{get_tf_data}
\alias{get_tf_data}
\title{Retrieves and formats data from the Tiny Forest Earthwatch website}
\usage{
get_tf_data()
}
\value{
A tidy data frame of Tiny Forest site information including GPS coordinates, planted date, area, and planted species for each site
}
\description{
This function scrapes the Tiny Forest Earthwatch website to obtain data about each Tiny Forest site including GPS coordinates, planted date, area, and planted species.
}
\examples{
get_tf_data()
}
