% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_tf.R
\name{initialize_tf}
\alias{initialize_tf}
\title{Initialize TinyForest package virtual environment}
\usage{
initialize_tf()
}
\value{
No explicit output, but this function creates and configures the required environment for TinyForest
}
\description{
This function initializes the TinyForest environment with required dependencies and libraries.
}
\examples{
initialize_tf()

}
