% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os_landcover.R
\name{plot_os_land_cover}
\alias{plot_os_land_cover}
\title{Plot OS Land Cover}
\usage{
plot_os_land_cover(lon, lat, key, buff = 1000, n = 5000L)
}
\arguments{
\item{lon}{numeric value indicating the longitude of the required point (e.g. -0.1426)}

\item{lat}{numeric value indicating the latitude of the required point (e.g. 51.5244)}

\item{key}{character value for the OS Data Hub API key in use}

\item{buff}{numeric value indicating the buffer around the specified point in metres (default = 1000)}

\item{n}{integer value indicating the maximum number of results to return (default = 10000)}
}
\value{
returns a list including the following objects:
\itemize{
\item\code{sf}: an sf object containing the queried data,
\item\code{map}: an interactive map generated using the queried data,
\item\code{tiers}: an sf object containing the tier B land cover information.}
}
\description{
Generate an interactive map of Ordnance Survey (OS) Land Cover data around a specified coordinate using the OS Data Hub API.
}
\details{
Requires the OS Data Hub API key and needed packages to run (tidyverse, reticulate, mapview, sf, tictoc, colourspace, leaflet).
}
\examples{

# Generate a plot of land cover around the London Eye
key <- "insert OS Data Hub API key"
plot_os_land_cover(lon = -0.1194, lat = 51.5033, key = key)

}
