% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_images.R
\name{save_ggmap_images}
\alias{save_ggmap_images}
\title{Save ggmap Images}
\usage{
save_ggmap_images(lon, lat, key)
}
\arguments{
\item{lon}{longitude of the center of the image}

\item{lat}{latitude of the center of the image}

\item{key}{Google API key for ggmap imagery}
}
\description{
This function saves a ggmap satellite image given a longitude and latitude.
}
\examples{
save_ggmap_images(-71.0636, 42.3581, "API_key")

}
