% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-data.R
\docType{data}
\name{tf_data}
\alias{tf_data}
\title{Tiny Forest (TF) data}
\format{
A data frame with 190 rows and 7 variables:
\describe{
\item{\code{tf_id}}{double COLUMN_DESCRIPTION}
\item{\code{plant_date}}{character COLUMN_DESCRIPTION}
\item{\code{area}}{double COLUMN_DESCRIPTION}
\item{\code{trees}}{list COLUMN_DESCRIPTION}
\item{\code{lat}}{double COLUMN_DESCRIPTION}
\item{\code{lon}}{double COLUMN_DESCRIPTION}
\item{\code{date}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
tf_data
}
\description{
TF ids, planting dates, area, lat-lon and tree species planted
}
\keyword{datasets}
