% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyForestR-package.R
\docType{package}
\name{tinyForestR-package}
\alias{tinyForestR}
\alias{tinyForestR-package}
\title{tinyForestR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julianflowers/tinyForestR}
  \item Report bugs at \url{https://github.com/julianflowers/tinyForestR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
