% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{calc_ndvi_buff}
\alias{calc_ndvi_buff}
\title{Calculate normalized difference vegetation index (NDVI) within a buffer around a point location}
\usage{
calc_ndvi_buff(
  lat = 25.1972,
  lon = 55.2744,
  dist = 500,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  cloud_cover = 10
)
}
\arguments{
\item{lat}{Numeric. The latitude of the point location of interest.}

\item{lon}{Numeric. The longitude of the point location of interest.}

\item{dist}{Numeric. The distance in meters to draw a buffer around the point location.}

\item{start_date}{Character. The start date in the format 'YYYY-MM-DD' for the time period of interest.}

\item{end_date}{Character. The end date in the format 'YYYY-MM-DD' for the time period of interest.}

\item{cloud_cover}{Numeric. The maximum allowable percentage of cloud cover in the satellite imagery used to calculate the NDVI. Default is 10.}
}
\value{
A list object containing the following components:
map: A map with two layers displaying the NDVI and the input point location.
tidy_dates: A tidied version of the image collection with start and end dates associated with each image in the collection#' image_ids: A vector of image IDs corresponding to the images used in the calculation.
raster: A raster layer object of the NDVI values within the buffer region.
ndvi_stats: Summary statistics of the NDVI values within the buffer region, including median, minimum, maximum, mean and standard deviation.
}
\description{
This function takes in a latitude, longitude, buffer distance, start and end time period, and cloud cover percentage,
and returns a map of NDVI within the buffer region, along with statistics on the NDVI values within the buffer.
}
\examples{
calc_ndvi_buff(lat = 25.1972, lon = 55.2744, dist = 500, start_date = '2021-01-01',
end_date = '2021-12-31',cloud_cover = 10)

}
\references{
\itemize{
\item COPERNICUS/S2_SR_HARMONIZED image collection.
\item rgee package documentation: https://csaybar.github.io/rgee/index.html
\item tidyrgee package documentation: https://csaybar.github.io/tidyrgee/index.html
\item zoo package documentation: https://cran.r-project.org/web/packages/zoo/index.html
\item terra package documentation: https://rspatial.org/terr'
}
}
