% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_data.R
\docType{data}
\name{tf_data}
\alias{tf_data}
\title{Tiny Forest data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 243 rows and 11 columns.
}
\usage{
tf_data
}
\description{
This dataset contains
}
\keyword{datasets}
