% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyForestR-package.R
\docType{package}
\name{tinyForestR-package}
\alias{tinyForestR}
\alias{tinyForestR-package}
\title{tinyForestR: Extract, manipulate and analyse data for Tiny Forests in the UK}
\description{
Extract, mainpulate and analyse biodiversity, ecological and environmental data for Tiny Forests and their surroundings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julianflowers/tinyForestR}
  \item \url{https://julianflowers.github.io/tinyForestR/}
  \item Report bugs at \url{https://github.com/julianflowers/tinyForestR/issues}
}

}
\author{
\strong{Maintainer}: Julian Flowers \email{jrf128@student.aru.ac.uk} (\href{https://orcid.org/0000-0001-5262-4526}{ORCID})

}
\keyword{internal}
