% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyForestR-package.R
\docType{package}
\name{tinyforestR-package}
\alias{tinyforestR}
\alias{tinyforestR-package}
\title{tinyforestR: tinyforestR is an R package which is designed to help extract landcover and diversity data from user-defined areas around Tiny Forest locations}
\description{
tinyforestR is an R package which is designed to help extract landcover and diversity data from user-defined areas around Tiny Forest locations. It creates a Tiny Forest database by scraping the EarthWatch website, extracts biodiversity data from the NBN atlas, and provides functions for extracting information from satellite images and OS landcover datasets. It can be loaded by running devtools::install_github("julianflowers12/tinyforestR").
}
\author{
\strong{Maintainer}: Julian Flowers \email{julian.flowers12@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
